requirements.txt
python-telegram-bot[ext]
Flask
pymongo
